+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+2.6.0 / 2019-08-06
+==================
+
+* Browser.text will return newlines on </br>
+* DefaultTimeout option on create browser
+* Injection scripts are now added through evaluate, so are not checked by CSP
+* Static config file removed
+* Opening new tabs will no longer reload the page
+
 2.5.1 / 2019-07-27
 ==================